♻️ refactor(provider): move providers/solana_wallet_auth.rs to providers/solana.rs
